pyproject.toml
[tool.poetry]
name = "bigbrother"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
disnake = "^2.4.0"
loguru = "^0.6.0"
taskipy = "^1.10.1"
aioredis = "^2.0.1"
fastapi = "^0.75.1"
ormar = "^0.11.0"
alembic = "^1.7.7"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
uvicorn = "^0.17.6"
orjson = "^3.6.7"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[tool.black]
line-length = 120

[tool.taskipy.tasks]
bot = "python -m src.bot"
api = "uvicorn src.api.__main__:app --host 0.0.0.0 --port 8080"
lint = "black . && isort --profile black ."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
